Allow linking a subset of the WASI preview1 surface
The host module table built all 42 implementations from a single
dictionary literal, so every one stayed referenced and no linker could
drop the unused ones. On a device with a couple hundred KB of RAM that
means paying for all of WASI to write to stdout.

Split the table into capability groups and represent a capability as a
value carrying its own registrar, rather than an option set: an option
set would need a dispatch function testing every case, which references
every registrar and defeats stripping just as thoroughly.

Guests import a fixed list regardless of what they call, so linking a
subset would otherwise fail to instantiate on an unused import. Fill the
gaps from a table of names and types that references no implementation,
controlled by stubUnlinked and on by default.

link(to:store:) keeps its previous behaviour: all capabilities, plus the
stubs that were previously always defined.
